requirements.txt
twine
motor>=2.5.1
aiohttp>=3.7,<4
aiohttp-session>=2.9
Bubot_Core>=4.0.0
aioredis